+++ api/src/main/java/btrplace/model/constraint/Online.java
@@ -67,7 +67,7 @@ public class Online extends SatConstraint {
     }
 
     @Override
-    public SatConstraintChecker getChecker() {
+    public SatConstraintChecker<Online> getChecker() {
         return new OnlineChecker(this);
     }