+++ api/src/main/java/btrplace/model/constraint/Overbook.java
@@ -124,7 +124,7 @@ public class Overbook extends SatConstraint {
     }
 
     @Override
-    public SatConstraintChecker getChecker() {
+    public SatConstraintChecker<Overbook> getChecker() {
         return new OverbookChecker(this);
     }