+++ api/src/main/java/btrplace/model/constraint/Quarantine.java
@@ -62,7 +62,7 @@ public class Quarantine extends SatConstraint {
     }
 
     @Override
-    public SatConstraintChecker getChecker() {
+    public SatConstraintChecker<Quarantine> getChecker() {
         return new QuarantineChecker(this);
     }