+++ CHANGES.md
@@ -10,6 +10,7 @@ version ?????????
 - ChocoReconfigurationAlgorithmParams renamed to Parameters
 - bug fixes
 - move to Java 8
+- new constraint NoDelay
 
 version 0.38.2 - 19 Jun 2014
 ----------------------